@@ -444,7 +444,7 @@ class Sneaker(private var context: Context) : View.OnClickListener, LifecycleObs
             sneakerView.startAnimation(AnimationUtils.loadAnimation(context, R.anim.popup_show))
             if (mAutoHide) {
                 val handler = Handler()
-                handler.removeCallbacks(null)
+                handler.removeCallbacks(Runnable { })
                 handler.postDelayed({
                     removeView(sneakerView)
                     mDismissListener?.onDismiss()
@@ -463,7 +463,7 @@ class Sneaker(private var context: Context) : View.OnClickListener, LifecycleObs
             sneakerView.startAnimation(AnimationUtils.loadAnimation(context, R.anim.popup_show))
             if (mAutoHide) {
                 val handler = Handler()
-                handler.removeCallbacks(null)
+                handler.removeCallbacks(Runnable { })
                 handler.postDelayed({
                     removeView(sneakerView)
                     mDismissListener?.onDismiss()
